Created N_QUEEN_PRINTING_SOLUTIONS.cpp
The file contains C++ code for solving the classic N Queen Problem.
The code calculates all positions and prints all of them!
CAUTION: MAKE SURE ONE HAS SUFFICIENT MAIN MEMORY AND SWAP MEMORY FOR PROCESSING THE CODE FOR N>=16
A system with 8Gb of RAM will work for n<=15. But for n>15, the system requirements will go on increasing!